+++ src/ofday.mli
@@ -1,2 +1,0 @@
-include Ofday_intf.Ofday with type underlying = float
-                          and module Span := Span