+++ src/span.mli
@@ -1,1 +1,0 @@
-include Span_intf.Span with type underlying = float